pyproject.toml
[tool.poetry]
name = "alyautdinov-ruslan-cw-1"
version = "0.1.0"
description = ""
authors = ["Аляутдинов Руслан <rusalyautdinov@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
pandas = "^2.2.2"
python-dotenv = "^1.0.1"
openpyxl = "^3.1.5"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.0"
mypy = "^1.11.0"
isort = "^5.13.2"
black = "^24.4.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.1"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'

[tool.black]
line-length = 119

[tool.isort]
line_length = 119